Move mocking to seperate file.
This will allow everything needed for mocking the reconcile loop
to be easily found in one place.
