+++ controllers/localvolumeset/status_test.go
@@ -7,40 +7,14 @@ import (
 
 	operatorv1 "github.com/openshift/api/operator/v1"
 	localv1alpha1 "github.com/openshift/local-storage-operator/api/v1alpha1"
-	"github.com/openshift/local-storage-operator/common"
 	"github.com/openshift/local-storage-operator/controllers/nodedaemon"
 	"github.com/stretchr/testify/assert"
 	appsv1 "k8s.io/api/apps/v1"
-	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
-	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 )
-
-const (
-	testNamespace = "default"
-)
-
-func newFakeLocalVolumeSetReconciler(t *testing.T, objs ...runtime.Object) *LocalVolumeSetReconciler {
-	scheme, err := localv1alpha1.SchemeBuilder.Build()
-	assert.NoErrorf(t, err, "creating scheme")
-
-	err = corev1.AddToScheme(scheme)
-	assert.NoErrorf(t, err, "adding corev1 to scheme")
-
-	err = appsv1.AddToScheme(scheme)
-	assert.NoErrorf(t, err, "adding appsv1 to scheme")
-
-	client := fake.NewFakeClientWithScheme(scheme, objs...)
-
-	return &LocalVolumeSetReconciler{
-		Client:   client,
-		Scheme:   scheme,
-		LvSetMap: &common.StorageClassOwnerMap{},
-	}
-}
 
 func TestDaemonSetCondition(t *testing.T) {
 	type knownResult struct {
